+++ src/adapters/index.ts
@@ -1,10 +1,14 @@
 import getArchiveOfOurOwnData from "./archiveofourown";
 import { getFFFavoritesData, getFFFollowingData } from "./fanfictiondotnet";
 import getQuestionableQuestingData from "./questionablequesting";
+import getSpaceBattlesData from "./spacebattles";
+import getSufficientVelocityData from "./sufficientvelocity";
 
 export {
   getArchiveOfOurOwnData,
   getFFFavoritesData,
   getFFFollowingData,
-  getQuestionableQuestingData
+  getQuestionableQuestingData,
+  getSpaceBattlesData,
+  getSufficientVelocityData
 };